Certificate: Fix type error when immediately handling certificates (Mantis 33660)

+++ Services/Certificate/classes/class.ilCertificateCron.php
@@ -311,7 +311,9 @@ class ilCertificateCron extends ilCronJob
             'usr_id: ' . $userId
         ]);
 
-        $this->queueRepository->removeFromQueue($entry->getId());
+        if ($entry->getId() !== null) {
+            $this->queueRepository->removeFromQueue($entry->getId());
+        }
 
         $this->dic->event()->raise(
             'Services/Certificate',
